Split q commands helper into services
